Compile built-in backends
The Win32 and OpenGL3 built-in ImGui backends are now compiled into the Rust library. Bindings are not generated for them so the user needs to create them manually, but this is the first step.
Revert update to latest cimgui and regenerated bindings

Note: do not update to latest cimgui version or regenerate the bindings on Windows due to this bug: https://github.com/rust-lang/rust-bindgen/issues/1966